# Conflicts:
#	src/test/java/gov/ca/cwds/rest/services/cms/AssignmentServiceTest.java
#	src/test/java/gov/ca/cwds/rest/services/cms/ReferralServiceTest.java

+++ src/main/java/gov/ca/cwds/data/persistence/cms/CwsOffice.java
@@ -18,7 +18,7 @@ public class CwsOffice extends CmsPersistentObject {
 
   @Basic
   @Column(name = "FAX_NO", nullable = false, precision = 0)
-  private int faxNumber;
+  private long faxNumber;
 
   @Basic
   @Column(name = "GEO_RGNTCD", nullable = false, length = 2)
@@ -42,7 +42,7 @@ public class CwsOffice extends CmsPersistentObject {
 
   @Basic
   @Column(name = "MSG_TEL_NO", nullable = false, precision = 0)
-  private int messagePhoneNumber;
+  private long messagePhoneNumber;
 
   @Basic
   @Column(name = "MSG_EXT_NO", nullable = false)
@@ -54,7 +54,7 @@ public class CwsOffice extends CmsPersistentObject {
 
   @Basic
   @Column(name = "PRM_TEL_NO", nullable = false, precision = 0)
-  private int primaryPhoneNumber;
+  private long primaryPhoneNumber;
 
   @Basic
   @Column(name = "PRM_EXT_NO", nullable = false)
@@ -99,10 +99,10 @@ public class CwsOffice extends CmsPersistentObject {
   public CwsOffice() {
   }
 
-  public CwsOffice(String id, int faxNumber, String geographicRegionCode, short governmentEntityType,
+  public CwsOffice(String id, long faxNumber, String geographicRegionCode, short governmentEntityType,
        String headquarterIndicator, String inactiveIndicator, String mailStopDescription,
-       int messagePhoneNumber, int messagePhoneExtensionNumber, String cwsOfficeNumber,
-       int primaryPhoneNumber, int primaryPhoneExtensionNumber, String fkStaffPerson,
+       long messagePhoneNumber, int messagePhoneExtensionNumber, String cwsOfficeNumber,
+       long primaryPhoneNumber, int primaryPhoneExtensionNumber, String fkStaffPerson,
        String commentDescription, String agencyName, String departmentDivisionName, String cwsOfficeName,
        String countySpecificCode, short agencyCodeNumber, short locationCountyType, String directorsNameTitle) {
     this.id = id;
@@ -136,11 +136,11 @@ public class CwsOffice extends CmsPersistentObject {
     this.id = identifier;
   }
 
-  public int getFaxNumber() {
+  public long getFaxNumber() {
     return faxNumber;
   }
 
-  public void setFaxNumber(int faxNo) {
+  public void setFaxNumber(long faxNo) {
     this.faxNumber = faxNo;
   }
 
@@ -184,11 +184,11 @@ public class CwsOffice extends CmsPersistentObject {
     this.mailStopDescription = mailstDsc;
   }
 
-  public int getMessagePhoneNumber() {
+  public long getMessagePhoneNumber() {
     return messagePhoneNumber;
   }
 
-  public void setMessagePhoneNumber(int msgTelNo) {
+  public void setMessagePhoneNumber(long msgTelNo) {
     this.messagePhoneNumber = msgTelNo;
   }
 
@@ -208,11 +208,11 @@ public class CwsOffice extends CmsPersistentObject {
     this.cwsOfficeNumber = cwsOffNo;
   }
 
-  public int getPrimaryPhoneNumber() {
+  public long getPrimaryPhoneNumber() {
     return primaryPhoneNumber;
   }
 
-  public void setPrimaryPhoneNumber(int prmTelNo) {
+  public void setPrimaryPhoneNumber(long prmTelNo) {
     this.primaryPhoneNumber = prmTelNo;
   }
